refactor(IND-246): extract fetchConnectToken to shared utility, fix misplaced JSDoc
- Move fetchConnectToken to src/lib/utils/connect-token.ts (was
  duplicated verbatim in both pollers)
- Fix daily-digest JSDoc that was displaced above fetchConnectToken
  instead of above handle()
