+++ packages/openclaw-plugin/src/polling/ambient-discovery/ambient-discovery.poller.ts
@@ -3,6 +3,7 @@ import { dispatchToMainAgent } from '../../lib/delivery/main-agent.dispatcher.js
 import { buildMainAgentPrompt } from '../../lib/delivery/main-agent.prompt.js';
 import { readMainAgentToolUse } from '../../lib/delivery/config.js';
 import { hashOpportunityBatch } from '../../lib/utils/hash.js';
+import { fetchConnectToken } from '../../lib/utils/connect-token.js';
 
 export interface AmbientDiscoveryConfig {
   baseUrl: string;
@@ -85,37 +86,6 @@ async function fetchAmbientDeliveredToday(
     api.logger.warn('Ambient stats fetch errored', {
       url,
       agentId: config.agentId,
-      error: err instanceof Error ? err.message : String(err),
-    });
-    return null;
-  }
-}
-
-/**
- * Mint a connect token for an opportunity via the backend.
- * Returns the token string, or null on failure (candidate will be skipped).
- */
-async function fetchConnectToken(
-  api: OpenClawPluginApi,
-  baseUrl: string,
-  apiKey: string,
-  opportunityId: string,
-): Promise<string | null> {
-  try {
-    const res = await fetch(`${baseUrl}/api/opportunities/${opportunityId}/connect-token`, {
-      method: 'POST',
-      headers: { 'x-api-key': apiKey },
-      signal: AbortSignal.timeout(10_000),
-    });
-    if (!res.ok) {
-      api.logger.warn('Connect token mint failed', { opportunityId, status: res.status });
-      return null;
-    }
-    const body = (await res.json()) as { token?: string };
-    return body.token ?? null;
-  } catch (err) {
-    api.logger.warn('Connect token mint errored', {
-      opportunityId,
       error: err instanceof Error ? err.message : String(err),
     });
     return null;
