+++ packages/openclaw-plugin/src/polling/daily-digest/daily-digest.poller.ts
@@ -3,6 +3,7 @@ import { dispatchToMainAgent } from '../../lib/delivery/main-agent.dispatcher.js
 import { buildMainAgentPrompt } from '../../lib/delivery/main-agent.prompt.js';
 import { readMainAgentToolUse } from '../../lib/delivery/config.js';
 import { hashOpportunityBatch } from '../../lib/utils/hash.js';
+import { fetchConnectToken } from '../../lib/utils/connect-token.js';
 
 export interface DailyDigestConfig {
   baseUrl: string;
@@ -21,37 +22,6 @@ const PENDING_LIMIT = 20;
  *
  * @returns true on successful dispatch, false on empty or error.
  */
-/**
- * Mint a connect token for an opportunity via the backend.
- * Returns the token string, or null on failure (candidate will be skipped).
- */
-async function fetchConnectToken(
-  api: OpenClawPluginApi,
-  baseUrl: string,
-  apiKey: string,
-  opportunityId: string,
-): Promise<string | null> {
-  try {
-    const res = await fetch(`${baseUrl}/api/opportunities/${opportunityId}/connect-token`, {
-      method: 'POST',
-      headers: { 'x-api-key': apiKey },
-      signal: AbortSignal.timeout(10_000),
-    });
-    if (!res.ok) {
-      api.logger.warn('Connect token mint failed', { opportunityId, status: res.status });
-      return null;
-    }
-    const body = (await res.json()) as { token?: string };
-    return body.token ?? null;
-  } catch (err) {
-    api.logger.warn('Connect token mint errored', {
-      opportunityId,
-      error: err instanceof Error ? err.message : String(err),
-    });
-    return null;
-  }
-}
-
 export async function handle(
   api: OpenClawPluginApi,
   config: DailyDigestConfig,
